refactor(LearningCircle):Interfaces to separate page
