+++ src/modules/Dashboard/modules/LearningCircle/services/LearningCircleInterface.ts
@@ -1,0 +1,11 @@
+export interface LcType {
+    circle_code: string;
+    created_by: string;
+    id: string;
+    ig: string;
+    meet_place: string;
+    meet_time: string;
+    member_count: number;
+    name: string;
+    org: string;
+}
